requirements.txt
flask
python-dotenv
google-generativeai
os
waitress
gunicorn